+++ Libraries/ExerciseVirtual.cpp
@@ -1,6 +1,6 @@
 #include "ExerciseVirtual.h"
 
-miit::algebra::ExerciseVirtual::ExerciseVirtual(int rows, int columns,Generator& generator)
+miit::algebra::ExerciseVirtual::ExerciseVirtual(size_t rows, size_t columns,Generator& generator)
 {
 	this->matrix = Matrix<int>{ rows, columns };
 	matrix.Fill(generator);
